Fix recipients by document field
When having multiple documents in the loop, email_by_document_field would be overridden by the mail address(es). In my case every trigger would only send one email, because for the doc following on the first one, the email_by_document_field would contain mail addresses instead of the field name.

Prefixing with tmp_ solves the problem, because we only need the key in that lines.