fix(doctype): Fieldname should not be in restricted fieldnames
If fieldname wasn't provided and the fieldname generated from label
was found to be in restricted names then 1 was appended to the newly
prepared fieldname
e.g. Name becomes name1

But the fieldname was allowed to be edited later to bypass this restriction
i.e.. name1 can be renamed to name

This fixes that.

Co-authored-by: Suraj Shetty <surajshetty3416@gmail.com>