fix: DO NOT ever delete site folder
this is very stupid, never delete folders that you do not know the
importance of

all the site data is stored in the sites folder, and may it even be
new-site, the data should not be deleted at any cost

please never add anything to the code that "deletes folders" on failure.
this is highly risky, and it should never be a practice to delete anything
without user confirmation.

Signed-off-by: Chinmay D. Pai <chinmaydpai@gmail.com>