fix: Login is failing when allow_login_using_mobile_number is set
When allow_login_using_mobile_number is set, we are doing auth
validation With respect to mobile_no with the assumption that
all logins use mobile number for validation and that is not really the case.

Fix is, if auth username does not match with any mobile number,
then go for username validation to find the user.