fix: refresh form section while refreshing the field
If all the fields with in a section has depends_on property and in such
case section itself stays hidden as all fields with in it are hidden.

Currently when any form field is updated, we are refreshing only the
fields but not the sections. So, field refresh is not affecting the form
because of section being hidden.

Fix is to refresh the sections when ever form fields are refreshed.

(cherry picked from commit 20c6e24a9f58a303b36da8cce247d273e6e3582f)