fix: test failure due to uncommited transaction
Steps to reproduce:
- Create a fresh new site with Frappe
- Add and install a new app.
- Add basic test_file.py anywhere in module.
- Run the single using test with bench command:
`bench --module module.dotted.path --test testname`

The test will not run and complain about implicit commit.

- root cause: test runner makes changes to DB but does not commit. e.g.
disabling of scheduler on L57.

Note: this is not reproducible in FF or when ERPNext is installed
because in many places of "test process" somethng else commits the
changes like `before_tests` hook, which usually isn't present in new
apps.