fix: Rollback, Release locals & Close db connection
When exitting console, rollback and destroy the Frappe connection. This
rollback is added so that rollback_observers are executed in case
methods are run in the console which shouldn't be committed.

For instance, running File.optimize will make changes to the file on
disk but may not update the corresponding data in the DB.