perf: Make request for filters_config once
Prior to this, a request was made on every filter change for a
filters_config entry. The options would not change based on the List /
filters context since no such data is shared with the API. Hence, it
can be stored upon first request for re-use.

Eg: Current usage in ERPNext - fiscal year operator's options don't change
frequently.