fix(email): Maintain multi-tenancy translation sanity in error messages
Defining translated strings as globals means the strings are evaluated
at import time. So the value of these variables is set based on which
site or user initializes the module. This makes it possible for  an
english user to get Russian errors.

---

Other changes: Remove unreachable code, avoid importing _socket.error alias to
OSError