fix: "You viewed/edited/created this" (#17365)
### Issues
In the form timeline we show events like “{user} {created/viewed/edited} this {some_time_ago}”. If user is the current user, we used “You” instead of “Jon Doe”. This works in English, but in other languages it messes up the rest of the sentence.

Because of this, people added wrong translations of “You” that made the grammar work, but also made the message less clear.

Also, the _last viewed_ message used to link to **View Log**. That can be read by system managers only and doesn't hold additional information. So it's just a source of errors for users trying to click on it.

### Fixes

- don't mix up users and salutations. Always show user's full name.
- change the _last viewed_ message to work like the other two, remove link to **View Log**

### Demo

Logged an as Administrator:

English

![Bildschirmfoto 2022-06-30 um 15 28 37](https://user-images.githubusercontent.com/14891507/176689772-dd878c90-4a2c-4a32-aa6c-d2e14a6340e1.png)

German

![Bildschirmfoto 2022-06-30 um 15 28 03](https://user-images.githubusercontent.com/14891507/176689812-ce13bb31-6c70-40f2-8699-6fb0db1d06f9.png)