perf: lazy load sidebar stats
Sidebar stats take a huge time to load on large tables and it's not
required everytime you load the page. Best to only trigger it when user
clicks on "tags" (or other stats filter)

This is exactly how assignment and other filters work, so no need for
sidebar stats to work differently.