fix: only update primary action if it is new action is provided
Currently, the primary action is updated even if the new action is None.
This breaks some old code that relies on the primary action being None
when no action is provided. (used to update primary action label)

I changed the code to only update the primary action
if the new action function is provided