fix(Build): reinit app_path, avoid get_app_path
frappe/esbuild seems like a ticking timebomb, I have
not seen so many global variables since 2nd year of
engineering.

The function get_app_path just returns a value from an
object that is initialized when the file is imported. Why
is it a function in the first place. What lawless idiocy
is this?