perf: Apply backpressure on link counts (#32712)
Context: This is QoL feature to highlight most used items. But in high
throughput environments where a lot of new documents are being created
this becomes a bottleneck.

Fix: Limit the size of counts that can be buffered before they're flushed.

Statistically this will still work just as well as it did before.